lsmac_qr_generator_chromebook/web_qr_generator_package/requirements.txt
Flask==2.3.3
qrcode[pil]==7.4.2
requests==2.31.0
Pillow==10.0.1 